requirements.txt
Django==2.2.5
arrow==0.15.2
urllib3==1.25.5
psycopg2==2.8.3
django-constance[database]==2.4.0
redis==3.3.8
django-redis==4.10.0
djangorestframework==3.10.3
celery==4.3.0
pymongo==3.9.0
bleach==3.1.0
django-cors-headers==3.2.1
